Added invert logic mode support for mod_loaded condition. Added invasion_rate condition for having things only work every x number of invasions. Added more debug output.

+++ src/main/java/CoroUtil/difficulty/data/spawns/DataActionMobSpawns.java
@@ -34,7 +34,7 @@ public class DataActionMobSpawns {
         return copy;
     }
 
-    public int getMaxMobCountDynamic(float difficulty) {
+    public int getMaxMobCountDynamic(double difficulty) {
         int count = this.count;
         count += count * (difficulty * count_difficulty_multiplier);
         if (count_max != -1 && count > count_max) {
